# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fitness tracking API with AI calorie detection
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Activity {
  id: string;
  steps: number;
  caloriesBurned: number;
  date: string;
  note?: string;
}

export interface CreateActivityInput {
  steps: number;
  caloriesBurned: number;
  date?: string;
  note?: string;
}

export interface ActivitySummary {
  todaySteps: number;
  todayCaloriesBurned: number;
  todayCaloriesConsumed: number;
  weeklySteps: number;
  weeklyCaloriesBurned: number;
  weeklyCaloriesConsumed: number;
  monthlySteps: number;
  recentActivities: Activity[];
}

export interface Goals {
  dailySteps: number;
  dailyCaloriesBurned: number;
  dailyCaloriesConsumed: number;
}

export interface FoodAnalyzeInput {
  imageBase64: string;
}

export interface FoodAnalysis {
  foodName: string;
  calories: number;
  confidence: string;
  protein?: number;
  carbs?: number;
  fat?: number;
  servingSize?: string;
}

export type LogFoodInputMealType =
  (typeof LogFoodInputMealType)[keyof typeof LogFoodInputMealType];

export const LogFoodInputMealType = {
  breakfast: "breakfast",
  lunch: "lunch",
  dinner: "dinner",
  snack: "snack",
} as const;

export interface LogFoodInput {
  foodName: string;
  calories: number;
  imageBase64?: string;
  protein?: number;
  carbs?: number;
  fat?: number;
  mealType?: LogFoodInputMealType;
}

export interface FoodLog {
  id: string;
  foodName: string;
  calories: number;
  imageBase64?: string;
  protein?: number;
  carbs?: number;
  fat?: number;
  mealType?: string;
  date: string;
}

export interface AdminLoginInput {
  username: string;
  password: string;
}

export interface AdminLoginResponse {
  token: string;
  message: string;
}

export interface AdminDashboard {
  totalActivities: number;
  totalFoodLogs: number;
  totalSteps: number;
  totalCaloriesBurned: number;
  totalCaloriesConsumed: number;
  recentActivities: Activity[];
  recentFoodLogs: FoodLog[];
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export type GetActivitiesParams = {
  period?: GetActivitiesPeriod;
};

export type GetActivitiesPeriod =
  (typeof GetActivitiesPeriod)[keyof typeof GetActivitiesPeriod];

export const GetActivitiesPeriod = {
  day: "day",
  week: "week",
  month: "month",
} as const;