# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fitness tracking API with AI calorie detection
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Add activity entry
 */
export const CreateActivityBody = zod.object({
  steps: zod.number(),
  caloriesBurned: zod.number(),
  date: zod.string().optional(),
  note: zod.string().optional(),
});

/**
 * @summary Get all activity entries
 */
export const GetActivitiesQueryParams = zod.object({
  period: zod.enum(["day", "week", "month"]).optional(),
});

export const GetActivitiesResponseItem = zod.object({
  id: zod.string(),
  steps: zod.number(),
  caloriesBurned: zod.number(),
  date: zod.string(),
  note: zod.string().optional(),
});
export const GetActivitiesResponse = zod.array(GetActivitiesResponseItem);

/**
 * @summary Get aggregated activity summary
 */
export const GetActivitySummaryResponse = zod.object({
  todaySteps: zod.number(),
  todayCaloriesBurned: zod.number(),
  todayCaloriesConsumed: zod.number(),
  weeklySteps: zod.number(),
  weeklyCaloriesBurned: zod.number(),
  weeklyCaloriesConsumed: zod.number(),
  monthlySteps: zod.number(),
  recentActivities: zod.array(
    zod.object({
      id: zod.string(),
      steps: zod.number(),
      caloriesBurned: zod.number(),
      date: zod.string(),
      note: zod.string().optional(),
    }),
  ),
});

/**
 * @summary Get user goals
 */
export const GetGoalsResponse = zod.object({
  dailySteps: zod.number(),
  dailyCaloriesBurned: zod.number(),
  dailyCaloriesConsumed: zod.number(),
});

/**
 * @summary Update user goals
 */
export const UpdateGoalsBody = zod.object({
  dailySteps: zod.number(),
  dailyCaloriesBurned: zod.number(),
  dailyCaloriesConsumed: zod.number(),
});

export const UpdateGoalsResponse = zod.object({
  dailySteps: zod.number(),
  dailyCaloriesBurned: zod.number(),
  dailyCaloriesConsumed: zod.number(),
});

/**
 * @summary Analyze food image with AI
 */
export const AnalyzeFoodBody = zod.object({
  imageBase64: zod.string(),
});

export const AnalyzeFoodResponse = zod.object({
  foodName: zod.string(),
  calories: zod.number(),
  confidence: zod.string(),
  protein: zod.number().optional(),
  carbs: zod.number().optional(),
  fat: zod.number().optional(),
  servingSize: zod.string().optional(),
});

/**
 * @summary Log a food item
 */
export const LogFoodBody = zod.object({
  foodName: zod.string(),
  calories: zod.number(),
  imageBase64: zod.string().optional(),
  protein: zod.number().optional(),
  carbs: zod.number().optional(),
  fat: zod.number().optional(),
  mealType: zod.enum(["breakfast", "lunch", "dinner", "snack"]).optional(),
});

/**
 * @summary Get food logs
 */
export const GetFoodLogsResponseItem = zod.object({
  id: zod.string(),
  foodName: zod.string(),
  calories: zod.number(),
  imageBase64: zod.string().optional(),
  protein: zod.number().optional(),
  carbs: zod.number().optional(),
  fat: zod.number().optional(),
  mealType: zod.string().optional(),
  date: zod.string(),
});
export const GetFoodLogsResponse = zod.array(GetFoodLogsResponseItem);

/**
 * @summary Admin login
 */
export const AdminLoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const AdminLoginResponse = zod.object({
  token: zod.string(),
  message: zod.string(),
});

/**
 * @summary Get admin dashboard data
 */
export const GetAdminDashboardResponse = zod.object({
  totalActivities: zod.number(),
  totalFoodLogs: zod.number(),
  totalSteps: zod.number(),
  totalCaloriesBurned: zod.number(),
  totalCaloriesConsumed: zod.number(),
  recentActivities: zod.array(
    zod.object({
      id: zod.string(),
      steps: zod.number(),
      caloriesBurned: zod.number(),
      date: zod.string(),
      note: zod.string().optional(),
    }),
  ),
  recentFoodLogs: zod.array(
    zod.object({
      id: zod.string(),
      foodName: zod.string(),
      calories: zod.number(),
      imageBase64: zod.string().optional(),
      protein: zod.number().optional(),
      carbs: zod.number().optional(),
      fat: zod.number().optional(),
      mealType: zod.string().optional(),
      date: zod.string(),
    }),
  ),
});

/**
 * @summary Get all food logs (admin)
 */
export const AdminGetFoodLogsResponseItem = zod.object({
  id: zod.string(),
  foodName: zod.string(),
  calories: zod.number(),
  imageBase64: zod.string().optional(),
  protein: zod.number().optional(),
  carbs: zod.number().optional(),
  fat: zod.number().optional(),
  mealType: zod.string().optional(),
  date: zod.string(),
});
export const AdminGetFoodLogsResponse = zod.array(AdminGetFoodLogsResponseItem);

/**
 * @summary Delete a food log (admin)
 */
export const AdminDeleteFoodLogParams = zod.object({
  id: zod.coerce.string(),
});

export const AdminDeleteFoodLogResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Get all activities (admin)
 */
export const AdminGetActivitiesResponseItem = zod.object({
  id: zod.string(),
  steps: zod.number(),
  caloriesBurned: zod.number(),
  date: zod.string(),
  note: zod.string().optional(),
});
export const AdminGetActivitiesResponse = zod.array(
  AdminGetActivitiesResponseItem,
);

/**
 * @summary Delete an activity (admin)
 */
export const AdminDeleteActivityParams = zod.object({
  id: zod.coerce.string(),
});

export const AdminDeleteActivityResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});