# lib/api-zod/src/generated/types/activity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fitness tracking API with AI calorie detection
 * OpenAPI spec version: 0.1.0
 */

export interface Activity {
  id: string;
  steps: number;
  caloriesBurned: number;
  date: string;
  note?: string;
}