# lib/api-zod/src/generated/types/activitySummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fitness tracking API with AI calorie detection
 * OpenAPI spec version: 0.1.0
 */
import type { Activity } from "./activity";

export interface ActivitySummary {
  todaySteps: number;
  todayCaloriesBurned: number;
  todayCaloriesConsumed: number;
  weeklySteps: number;
  weeklyCaloriesBurned: number;
  weeklyCaloriesConsumed: number;
  monthlySteps: number;
  recentActivities: Activity[];
}