# lib/api-zod/src/generated/types/adminDashboard.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fitness tracking API with AI calorie detection
 * OpenAPI spec version: 0.1.0
 */
import type { Activity } from "./activity";
import type { FoodLog } from "./foodLog";

export interface AdminDashboard {
  totalActivities: number;
  totalFoodLogs: number;
  totalSteps: number;
  totalCaloriesBurned: number;
  totalCaloriesConsumed: number;
  recentActivities: Activity[];
  recentFoodLogs: FoodLog[];
}