# lib/api-zod/src/generated/types/foodAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fitness tracking API with AI calorie detection
 * OpenAPI spec version: 0.1.0
 */

export interface FoodAnalysis {
  foodName: string;
  calories: number;
  confidence: string;
  protein?: number;
  carbs?: number;
  fat?: number;
  servingSize?: string;
}