# lib/api-zod/src/generated/types/foodAnalyzeInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fitness tracking API with AI calorie detection
 * OpenAPI spec version: 0.1.0
 */

export interface FoodAnalyzeInput {
  imageBase64: string;
}