# lib/api-zod/src/generated/types/foodLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fitness tracking API with AI calorie detection
 * OpenAPI spec version: 0.1.0
 */

export interface FoodLog {
  id: string;
  foodName: string;
  calories: number;
  imageBase64?: string;
  protein?: number;
  carbs?: number;
  fat?: number;
  mealType?: string;
  date: string;
}